v2.0

+++ samples/sample_prime_numbers.cpp
@@ -16,7 +16,7 @@
 
 int main()
 {
-  long long n, m, k, count;
+  int n, m, k, count;
 
   setlocale(LC_ALL, "Russian");
   cout << "Тестирование программ поддержки битового поля" << endl;
